+++ src/platform/mod.rs
@@ -724,9 +724,6 @@ mod platform_impl {
         )))
     )]
     pub use crate::host::jack::Host as JackHost;
-    #[cfg(feature = "pulseaudio")]
-    pub use crate::host::pulseaudio::Host as PulseAudioHost;
-
     #[cfg(feature = "pipewire")]
     #[cfg_attr(
         docsrs,
@@ -741,16 +738,42 @@ mod platform_impl {
         )))
     )]
     pub use crate::host::pipewire::Host as PipeWireHost;
+    #[cfg(feature = "pulseaudio")]
+    #[cfg_attr(
+        docsrs,
+        doc(cfg(all(
+            any(
+                target_os = "linux",
+                target_os = "dragonfly",
+                target_os = "freebsd",
+                target_os = "netbsd"
+            ),
+            feature = "pulseaudio"
+        )))
+    )]
+    pub use crate::host::pulseaudio::Host as PulseAudioHost;
     impl_platform_host!(
+        #[cfg(feature = "pipewire")] PipeWire => PipeWireHost,
         #[cfg(feature = "pulseaudio")] PulseAudio => PulseAudioHost,
         #[cfg(feature = "jack")] Jack => JackHost,
         Alsa => AlsaHost,
         #[cfg(feature = "custom")] Custom => super::CustomHost,
-        #[cfg(feature = "pipewire")] PipeWire => super::PipeWireHost,
     );
 
     /// The default host for the current compilation target platform.
     pub fn default_host() -> Host {
+        #[cfg(feature = "pipewire")]
+        if <PipeWireHost as crate::traits::HostTrait>::is_available() {
+            if let Ok(host) = PipeWireHost::new() {
+                return host.into();
+            }
+        }
+        #[cfg(feature = "pulseaudio")]
+        if <PulseAudioHost as crate::traits::HostTrait>::is_available() {
+            if let Ok(host) = PulseAudioHost::new() {
+                return host.into();
+            }
+        }
         AlsaHost::new()
             .expect("the default host should always be available")
             .into()